Added unit tests
particle_species_sets and it's mods were removed and an inlined copy of game_of_life.rs was added to the tests mod as it will only be used in testing.
Code that was in main() to test the creation of an empty World has been moved to the tests mod as a test.